Rename broadcasting.py to composite.py
