+++ baybe/surrogates/__init__.py
@@ -1,7 +1,7 @@
 """BayBE surrogates."""
 
 from baybe.surrogates.bandit import BetaBernoulliMultiArmedBanditSurrogate
-from baybe.surrogates.broadcasting import BroadcastingSurrogate
+from baybe.surrogates.composite import BroadcastingSurrogate
 from baybe.surrogates.custom import CustomONNXSurrogate, register_custom_architecture
 from baybe.surrogates.gaussian_process.core import GaussianProcessSurrogate
 from baybe.surrogates.linear import BayesianLinearSurrogate
