+++ baybe/surrogates/base.py
@@ -41,7 +41,7 @@ if TYPE_CHECKING:
     from botorch.posteriors import GPyTorchPosterior, Posterior
     from torch import Tensor
 
-    from baybe.surrogates.broadcasting import BroadcastingSurrogate
+    from baybe.surrogates.composite import BroadcastingSurrogate
 
 _ONNX_ENCODING = "latin-1"
 """Constant signifying the encoding for onnx byte strings in pretrained models.
@@ -153,7 +153,7 @@ class Surrogate(ABC, SurrogateProtocol, SerialMixin):
         effectively disables the model's inherent multi-target mechanism by treating
         it as a single-target surrogate and applying the same broadcasting mechanism.
         """
-        from baybe.surrogates.broadcasting import BroadcastingSurrogate
+        from baybe.surrogates.composite import BroadcastingSurrogate
 
         return BroadcastingSurrogate(self)
