+++ baybe/surrogates/composite.py
